fix, editPoster & css

+++ js/pages/poster.js
@@ -1,15 +1,4 @@
-import {
-  doc,
-  getDoc,
-  getDocs,
-  collection,
-  query,
-  where,
-  deleteDoc,
-  updateDoc,
-  setDoc,
-  orderBy,
-} from "https://www.gstatic.com/firebasejs/9.14.0/firebase-firestore.js";
+import { doc, getDoc, getDocs, collection, query, where, deleteDoc, updateDoc, setDoc, orderBy } from "https://www.gstatic.com/firebasejs/9.14.0/firebase-firestore.js";
 import { authService, dbService } from "../firebase.js";
 import { getYYYYMMDD } from "../util.js";
 
@@ -45,21 +34,12 @@ export const getPosterInfo = async () => {
       const uid = authService.currentUser?.uid || testUid; //test
       const userProfileImage = getUserProfile(uid).profileImage;
 
-      if (userProfileImage)
-        document.getElementById("comment-user-img").src = userProfileImage;
-      if (image)
-        document.getElementById(
-          "post-img"
-        ).style.backgroundImage = `url(${image})`;
-      if (nickName)
-        document.getElementById("post-nickname").textContent = nickName;
-      if (babyName)
-        document.getElementById("post-animal-name").innerHTML = babyName;
-      if (createdAt)
-        document.getElementById("post-date").textContent =
-          getYYYYMMDD(createdAt);
-      if (profileImage)
-        document.getElementById("post-user-img").src = profileImage;
+      if (userProfileImage) document.getElementById("comment-user-img").src = userProfileImage;
+      if (image) document.getElementById("post-img").style.backgroundImage = `url(${image})`;
+      if (nickName) document.getElementById("post-nickname").textContent = nickName;
+      if (babyName) document.getElementById("post-animal-name").innerHTML = babyName;
+      if (createdAt) document.getElementById("post-date").textContent = getYYYYMMDD(createdAt);
+      if (profileImage) document.getElementById("post-user-img").src = profileImage;
       if (title) document.getElementById("post-title").innerHTML = title;
       if (content) document.getElementById("post-desc").innerHTML = content;
 
@@ -68,8 +48,8 @@ export const getPosterInfo = async () => {
         if (btnElement.children.length < 2) {
           const div = document.createElement("div");
           div.id = "post-btns";
-          const temp_html = `<img class="comment-btn" src="../assets/edit.png" width="36" height="36" />
-                              <img class="comment-btn" src="../assets/delete.png" width="36" height="36" />`;
+          const temp_html = `<img class="comment-btn" onclick="updatePoster();" src="../assets/edit.png" width="24" height="24" />
+                              <img class="comment-btn" onclick="deletePoster();" src="../assets/delete.png" width="24" height="24" />`;
           div.innerHTML = temp_html;
           btnElement.appendChild(div);
         }
@@ -87,16 +67,21 @@ export const getPosterInfo = async () => {
 export const updatePoster = () => {
   const docId = sessionStorage.getItem("docId");
   sessionStorage.setItem("v2", docId);
-  location.hash("#post-edit");
+  window.location.hash = "posting-edit";
 };
 
 export const deletePoster = async () => {
+  if (!window.confirm("게시물을 삭제하시겠습니까?")) {
+    return;
+  }
+
   const docId = sessionStorage.getItem("docId");
   if (!docId) return alert("다시 시도해주세요.");
 
   try {
     await deleteDoc(doc(dbService, "post", docId));
-    return alert("게시글을 삭제하였습니다.");
+    alert("게시글을 삭제하였습니다.");
+    return history.back();
   } catch (err) {
     console.error(err);
     return alert("다시 시도해주세요.");
@@ -121,23 +106,21 @@ export const getCommentList = async () => {
       const { userId, postId, content, createdAt } = doc.data();
       const { profileImage, nickName } = await getUserProfile(userId);
       const temp_html = `<div class="comment-wrapper">
-                          <img class="comment-profile" src="${profileImage}" />
-                          <div class="comment-items">
-                            <div class="comment-header">
-                              <div class="comment-info">
-                                <div class="comment-nickname">${nickName}</div>
-                                <div class="comment-date">${getYYYYMMDD(
-                                  createdAt
-                                )}</div>
+                            <img class="comment-profile" src="${profileImage}" />
+                            <div class="comment-items">
+                              <div class="comment-header">
+                                <div class="comment-info">
+                                  <div class="comment-nickname">${nickName}</div>
+                                  <div class="comment-date">${getYYYYMMDD(createdAt)}</div>
+                                </div>
+                                <div class="comment-btns">
+                                  <img class="comment-btn" onclick="editComment('${commentId}');" src="../assets/edit.png" width="24" height="24" />
+                                  <img class="comment-btn" onclick="deleteComment('${commentId}');" src="../assets/delete.png" width="24" height="24" />
+                                </div>
                               </div>
-                              <div class="comment-btns">
-                                <img class="comment-btn" onclick="editComment('${commentId}');" src="../assets/edit.png" width="36" height="36" />
-                                <img class="comment-btn" onclick="deleteComment('${commentId}');" src="../assets/delete.png" width="36" height="36" />
-                              </div>
+                              <div class="comment-contents">${content}</div>
                             </div>
-                            <div class="comment-contents">${content}</div>
-                          </div>
-                        </div>`;
+                          </div>`;
 
       const div = document.createElement("div");
       div.innerHTML = temp_html;
@@ -182,8 +165,7 @@ export const editComment = (commentId) => {
   parent.appendChild(btnsElement);
 
   const parent_2 = parent.parentNode;
-  const contentsElement =
-    parent_2.getElementsByClassName("comment-contents")[0];
+  const contentsElement = parent_2.getElementsByClassName("comment-contents")[0];
   const value = contentsElement.innerHTML.trim();
 
   contentsElement.remove();
@@ -199,9 +181,7 @@ export const cancelEditComment = () => {
 
 export const updateComment = async (commentId) => {
   const containerElement = window.event.target.parentNode.parentNode.parentNode;
-  const content = containerElement.getElementsByClassName(
-    "comment-contents-edit"
-  )[0].value;
+  const content = containerElement.getElementsByClassName("comment-contents-edit")[0].value;
 
   if (!commentId || !content) return alert("다시 시도해주세요.");
